pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "clawinterview"
version = "0.1.0"
description = "Typed, reusable intake layer for orchestrated ClawPipe pipelines"
requires-python = ">=3.11"
dependencies = [
    "PyYAML>=6.0",
    "jsonschema>=4.21",
    "pydantic>=2.0",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
clawinterview = "clawinterview.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/clawinterview"]
